+++ src/card/cards/stale_followups.rs
@@ -66,3 +66,5 @@ pub const SHIMIAN_NIGHT_STALKER: CardDefinitionId = CardDefinitionId(1691);
 pub const MURK_DWELLERS: CardDefinitionId = CardDefinitionId(1711);
 pub const THE_WRETCHED: CardDefinitionId = CardDefinitionId(1712);
 pub const ORGG: CardDefinitionId = CardDefinitionId(1713);
+pub const SIMULACRUM: CardDefinitionId = CardDefinitionId(1714);
+pub const REVERSE_POLARITY: CardDefinitionId = CardDefinitionId(1715);